# design/api.yaml
openapi: 3.0.3
info:
  title: Gridow API
  description: 格物（Gridow）物品管理后台服务接口文档，本地数据作为缓存，后台服务作为数据提供者。
  version: 1.0.0
  contact:
    name: Gridow Team
servers:
  - url: https://api.gridow.com/v1
    description: 生产环境
  - url: http://localhost:8080/v1
    description: 开发环境
...
paths:
...
  /items/{itemId}:
    get:
      tags: [物品]
      summary: 获取物品详情
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [物品]
      summary: 更新物品
      description: 编辑物品信息，若位置变更则自动更新空间关联和计数
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdateRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [物品]
      summary: 删除物品
      description: 删除物品，同时删除空间关联和历史记录，更新空间计数
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: 删除成功
        "404":
          $ref: "#/components/responses/NotFound"
...
  /history/item/{itemId}:
    get:
      tags: [操作历史]
      summary: 获取物品的操作历史
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryRecord"
...
components:
  parameters:
    ItemId:
      name: itemId
      in: path
      required: true
      description: 物品ID
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 错误码
        message:
          type: string
          description: 错误信息
...
    Item:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          description: 物品ID
        name:
          type: string
          description: 物品名称
          example: 洗衣液
        icon:
          type: string
          description: 物品图标（Emoji）
          default: "📦"
          example: "🧴"
        qty:
          type: integer
          description: 库存数量
          default: 0
          example: 3
        location:
          type: string
          description: 所在位置名称
          example: "🏠 客厅 > 🗄️ 电视柜"
        locationId:
          type: string
          format: uuid
          description: 所在空间ID
          nullable: true
        category:
          type: string
          description: 物品分类
          enum: [daily, food, tool, medicine, clothes, electronics]
          default: daily
          example: daily
        tags:
          type: string
          description: 标签（JSON数组字符串）
          default: "[]"
          example: '["清洁","日用品"]'
        barcode:
          type: string
          description: 条码/二维码内容
          default: ""
          example: "6901234567890"
        photos:
          type: string
          description: 照片URL列表（JSON数组字符串）
          default: "[]"
          example: '["https://api.gridow.com/v1/images/abc123"]'
        photoUri:
          type: string
          description: 主图URL
          default: ""
          example: "https://api.gridow.com/v1/images/abc123"
        buyDate:
          type: string
          description: 购买日期
          default: ""
          example: "2025-01-15"
        expiry:
          type: string
          description: 过期日期，"-"表示无过期
          default: "-"
          example: "2026-01-15"
        remark:
          type: string
          description: 备注
          default: ""
        trackLowStock:
          type: boolean
          description: 是否跟踪低库存
          default: false
        ownerId:
          type: string
          description: 所有者用户ID
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
...
    ItemUpdateRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        qty:
          type: integer
          minimum: 0
        locationId:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          enum: [daily, food, tool, medicine, clothes, electronics]
        tags:
          type: array
          items:
            type: string
        barcode:
          type: string
        photoUri:
          type: string
        buyDate:
          type: string
        expiry:
          type: string
        remark:
          type: string
        trackLowStock:
          type: boolean
          description: 是否跟踪低库存
...
    HistoryRecord:
      type: object
      required: [id, type, itemId, itemName, qty, time]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          description: 操作类型
          enum: [in, out, move]
          example: in
        itemId:
          type: string
          format: uuid
          description: 物品ID
        itemName:
          type: string
          description: 物品名称
          example: 洗衣液
        qty:
          type: integer
          description: 操作数量
          example: 2
        fromLocation:
          type: string
          description: 来源位置（出库/转移时有值）
          example: "🏠 客厅"
        toLocation:
          type: string
          description: 目标位置（入库/转移时有值）
          example: "🏠 卧室 > 🗄️ 衣柜"
        reason:
          type: string
          description: 原因（出库时有值）
        remark:
          type: string
          description: 备注
        time:
          type: string
          description: 操作时间
          example: "2025-05-06 14:30"
...